re-organized backend according to resources. created a repo controller and repo service. trimmed the GitHub rest client down to only making rest calls factored the project entity controller and service out of the project ones.
