+++ src/main/java/dev/functionalnotpretty/githubpoc/githubclient/GithubRestClient.java
@@ -1,8 +1,6 @@
-package dev.functionalnotpretty.githubpoc.restservice;
+package dev.functionalnotpretty.githubpoc.githubclient;
 
-import dev.functionalnotpretty.githubpoc.entities.*;
 import dev.functionalnotpretty.githubpoc.exceptions.GitRequestException;
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.core.ParameterizedTypeReference;
@@ -10,19 +8,18 @@ import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClient;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
 public class GithubRestClient {
     private final static Logger log = LoggerFactory.getLogger(GithubRestClient.class);
     private final RestClient githubRestClient;
-    private final static String GITHUB_BASE_URL = "https://api.github.com";
 
+    private final static String GITHUB_BASE_URL = "https://api.github.com";
     private final static String token = System.getenv("GITHUB_TOKEN");
 
     public final static String GITHUB_LINK_HEADER = "Link";
-    private final static String REL_NEXT = "rel=\"next\"";
+    public final static String REL_NEXT = "rel=\"next\"";
 
     public GithubRestClient(RestClient.Builder restClientBuilder) {
         this.githubRestClient = restClientBuilder.baseUrl(GITHUB_BASE_URL)
@@ -30,75 +27,37 @@ public class GithubRestClient {
                 .defaultHeader("Accept", "application/vnd.github+json")
                 .defaultHeader("User-Agent", "amfritz")
                 .defaultHeader("X-GitHub-Api-Version", "2022-11-28")
-//                .defaultStatusHandler(HttpStatusCode::isError,  (req, resp) -> {
-//                    log.error("default github error handler, returned HTTP error code {} : {}",resp.getStatusCode(),resp.getBody());
-//                    throw new RuntimeException("a github error has occurred: " + resp.getStatusCode());
-//                })
-//                .requestInterceptor((request, body, execution) -> {
-//                    logRequest(request, body);
-//                    var response = execution.execute(request, body);
-//                    logResponse(request, response);
-//                    return response;
-//                })
                 .build();
     }
 
     // curl -i -L -H "Accept: application/vnd.github+json" -H "Authorization: Bearer " -H "X-GitHub-Api-Version: 2022-11-28"  https://api.github.com/user/repos
-    public List<GitHubRepo> getUserRepos() {
-        log.info("getUserRepos called");
-        return this.githubRestClient.get()
+    public ResponseEntity<List<GitHubRepo>> getUserRepos() {
+//        log.info("getUserRepos called");
+        return this.githubRestClient
+                .get()
                 .uri("/user/repos")
                 .retrieve()
-                .body(new ParameterizedTypeReference<>() {
+                .toEntity(new ParameterizedTypeReference<>() {
                 });
     }
 
     // // curl -i -L -H "Accept: application/vnd.github+json" -H "Authorization: Bearer " -H "X-GitHub-Api-Version: 2022-11-28"  https://api.github.com/repos/amfritz/my-github-project-poc/commits
     //
-    public List<GitHubRepoCommit> getUserRepoCommits(String userId, String repo) {
-        log.info("getUserRepoCommits called");
-
-        var result = new ArrayList<GitHubRepoCommit>();
+    public ResponseEntity<List<GitHubRepoCommit>>  getRepoCommitsByUser(String userId, String repo) {
         String uri = "/repos/{userid}/{repo}/commits";
+        return this.githubRestClient.get()
+                .uri(uri, userId, repo)
+                .retrieve()
+                .toEntity(new ParameterizedTypeReference<>() {
+                });
+    }
 
-        do {
-            ResponseEntity<List<GitHubRepoCommit>> resp = this.githubRestClient.get()
-                    .uri(uri, userId, repo)
-                    .retrieve()
-                    .toEntity(new ParameterizedTypeReference<>() {
-                    });
-
-            var commits = resp.getBody();
-            result.addAll(commits);
-
-            // check headers to see if there's a link for more commits. if so, iterate though the pages
-            if (resp.getHeaders().containsKey(GITHUB_LINK_HEADER)) {
-                String next = null;
-                String[] pageLinks = StringUtils.split(resp.getHeaders().get(GITHUB_LINK_HEADER).getFirst(), ',');
-                for (String s : pageLinks) {
-                    // the links could have rel="first", rel="prev", rel="next", rel="last" sent. only get the next one.
-                    // if no next then we are at the end and break out.
-                    String[] split = StringUtils.split(s, ';');
-                    if (StringUtils.trim(split[1]).equalsIgnoreCase(REL_NEXT)) {
-                        log.info("next link found: {}", split[0]);
-                        next = split[0];
-                    }
-                }
-                if (next ==  null) {
-                    log.info("next link not found, breaking out");
-                    break;
-                }
-                // the url is bracketed by < and > so strip them off
-                uri = StringUtils.strip( next, "<>");
-                log.info("getUserRepoCommits link [{}]", uri);
-            } else {
-                // no link header in response
-                break;
-            }
-        }
-        while(true);
-
-        return result;
+    public ResponseEntity<List<GitHubRepoCommit>>  getRepoCommitsPaginated(String uri) {
+        return this.githubRestClient.get()
+                .uri(uri)
+                .retrieve()
+                .toEntity(new ParameterizedTypeReference<>() {
+                });
     }
 
     public GitHubCreateWebhookResponse createGitHubRepositoryWebHook(String userId, String repository) {
