+++ src/main/java/dev/functionalnotpretty/githubpoc/project/ProjectDto.java
@@ -1,7 +1,5 @@
-package dev.functionalnotpretty.githubpoc.models;
+package dev.functionalnotpretty.githubpoc.project;
 
-import dev.functionalnotpretty.githubpoc.entities.ProjectRepo;
-import dev.functionalnotpretty.githubpoc.entities.ProjectStatus;
 import jakarta.validation.constraints.NotEmpty;
 
 public record ProjectDto(
