+++ src/main/java/dev/functionalnotpretty/githubpoc/project/ProjectRepo.java
@@ -1,4 +1,4 @@
-package dev.functionalnotpretty.githubpoc.entities;
+package dev.functionalnotpretty.githubpoc.project;
 
 public record ProjectRepo(
         String id,
