+++ src/main/java/dev/functionalnotpretty/githubpoc/webhooks/GitHubWebHook.java
@@ -1,15 +1,10 @@
-package dev.functionalnotpretty.githubpoc.controllers.webhooks;
+package dev.functionalnotpretty.githubpoc.webhooks;
 
-import com.azure.core.annotation.Headers;
-import dev.functionalnotpretty.githubpoc.controllers.ProjectsController;
-import dev.functionalnotpretty.githubpoc.services.GitHubWebhookService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.Map;
 
 @RestController
 @RequestMapping("api/webhooks/github")
@@ -23,10 +18,7 @@ public class GitHubWebHook {
 
     @PostMapping
     public ResponseEntity<String> doGitHubWebhook(@RequestBody String body, @RequestHeader HttpHeaders headers) {
-        log.info("received request: {}", body);
-//        for(var  entry : headers.entrySet()) {
-//            log.info("header, {}: {}", entry.getKey(), entry.getValue());
-//        }
+        log.trace("doGitHubWebhook called");
         this.gitHubWebhookService.handleWebhookRequest(headers, body);
         return ResponseEntity.accepted().body("");
     }
